Never skip custom CODEREF entries in @INC
When the last Module::Case sensitive module completes,
the special Module::Case CODEREF in @INC is removed for
efficiency purposes.
But Perl just gleefully continues onto the next @INC
entry based on index. So when an entry is removed from
@INC on the fly, then entry immediately following what
was removed will always be skipped.

To avoid exploiting this problem, now the $inc_sniffer
is injected right before the FIRST regular directory
entry instead of lazily tossing it at the very front.

This patch also fixes t/21-inc-ref-behavior-with.t